Fluent use of accessor methods

The 'proDAFit' object overwrites the dollar function to make it easy to call functions to access values inside the object. This has the advantage that it is very easy to discover the relevant methods but nonetheless have an isolated implementation. Unlike the `@` operator which directly accesses the underlying implementation, the `$` operator only exposes a limited set of functions

abundances

hyper_parameters

feature_parameters

coefficients

convergence

design

reference_level

result_names

coefficient_variance_matrices

colData

rowData

# S3 method for proDAFit
.DollarNames(x, pattern = "")

# S4 method for proDAFit
$(x, name)

# S4 method for proDAFit
$(x, name) <- value

Arguments

x
an object of class 'proDAFit' produced by proDA()

pattern
the regex pattern that is provided by the IDE

name
one of the functions listed above

value
Warning: modifying the content of a 'proDAFit' object is not allowed

Value

whatever the function called name returns.

See also

Examples

  syn_data <- generate_synthetic_data(n_proteins = 10)
  fit <- proDA(syn_data$Y, design = syn_data$groups)

  # The two styles are identical
  design(fit)

#>               Condition_1 Condition_2
#> Condition_1-1           1           0
#> Condition_1-2           1           0
#> Condition_1-3           1           0
#> Condition_2-1           0           1
#> Condition_2-2           0           1
#> Condition_2-3           0           1
#> attr(,"assign")
#> [1] 1 1
#> attr(,"contrasts")
#> attr(,"contrasts")$x_
#> [1] "contr.treatment"
#> 

  fit$design

#>               Condition_1 Condition_2
#> Condition_1-1           1           0
#> Condition_1-2           1           0
#> Condition_1-3           1           0
#> Condition_2-1           0           1
#> Condition_2-2           0           1
#> Condition_2-3           0           1
#> attr(,"assign")
#> [1] 1 1
#> attr(,"contrasts")
#> attr(,"contrasts")$x_
#> [1] "contr.treatment"
#> 

  # More functions
  fit$abundances

#>            Condition_1-1 Condition_1-2 Condition_1-3 Condition_2-1
#> protein_1       24.07919      23.42556      23.94491      24.21937
#> protein_2       18.61085      18.18919            NA            NA
#> protein_3             NA            NA            NA            NA
#> protein_4             NA      18.35705      17.45826            NA
#> protein_5       21.94793      20.08355      24.01483      21.14165
#> protein_6       23.71996      24.34934      24.07880      23.33419
#> protein_7             NA            NA      17.72383      18.07316
#> protein_8       21.88006      21.28270      21.53463      21.96805
#> protein_9       20.88883      21.35111      21.41965      21.15165
#> protein_10            NA            NA            NA            NA
#>            Condition_2-2 Condition_2-3
#> protein_1       23.50432      24.27628
#> protein_2       18.65685            NA
#> protein_3             NA            NA
#> protein_4             NA            NA
#> protein_5       21.54833      21.39306
#> protein_6       24.23279      23.77613
#> protein_7             NA            NA
#> protein_8       21.56534      21.77661
#> protein_9       21.78433      21.05557
#> protein_10            NA            NA