Column wise median normalization of the data matrix

The method calculates for each sample the median change (i.e. the difference between the observed value and the row average) and substracts it from each row. Missing values are ignored in the procedure. The method is based on the assumption that a majority of the rows did not change.

median_normalization(X)

Arguments

X
a matrix or SummarizedExperiment of proteins and samples

Value

the normalized matrix

Examples

  syn_data <- generate_synthetic_data(n_proteins = 10)
  normalized_data <- median_normalization(syn_data$Y)
  normalized_data

#>            Condition_1-1 Condition_1-2 Condition_1-3 Condition_2-1
#> protein_1       18.27498      18.47614      18.13454            NA
#> protein_2       26.61815      26.22856      26.32454      26.18305
#> protein_3             NA            NA            NA      17.87249
#> protein_4       18.57748            NA      18.23686      18.82792
#> protein_5       19.51317      19.48286      19.81974      19.28272
#> protein_6       21.13594      21.40975      21.77695      21.69377
#> protein_7             NA      20.08564      19.68163      19.66199
#> protein_8       22.33905      22.19973      22.32075      22.37467
#> protein_9       20.34292      19.91314            NA      20.57679
#> protein_10      19.48041      20.05136      20.06794      17.69139
#>            Condition_2-2 Condition_2-3
#> protein_1             NA            NA
#> protein_2       25.89577      25.83017
#> protein_3             NA            NA
#> protein_4             NA            NA
#> protein_5       19.39447      19.65915
#> protein_6       21.43720      21.48371
#> protein_7       19.91595      19.83197
#> protein_8       22.96830      22.70942
#> protein_9       20.30399      20.08436
#> protein_10            NA            NA