Row-wise tests of difference using the probabilistic dropout model

This is a helper function that combines the call of proDA() and test_diff(). If you need more flexibility use those functions.

pd_row_t_test(X, Y, moderate_location = TRUE, moderate_variance = TRUE,
  alternative = c("two.sided", "greater", "less"),
  pval_adjust_method = "BH", location_prior_df = 3, max_iter = 20,
  epsilon = 0.001, return_fit = FALSE, verbose = FALSE)

pd_row_f_test(X, ..., groups = NULL, moderate_location = TRUE,
  moderate_variance = TRUE, pval_adjust_method = "BH",
  location_prior_df = 3, max_iter = 20, epsilon = 0.001,
  return_fit = FALSE, verbose = FALSE)

Arguments

X, Y, ...
the matrices for condition 1, 2 and so on. They must have the same number of rows.

moderate_location
boolean values to indicate if the location and the variances are moderated. Default: TRUE

moderate_variance
boolean values to indicate if the location and the variances are moderated. Default: TRUE

alternative
a string that decides how the hypothesis test is done. This parameter is only relevant for the t-test / contrast test. Default: "two.sided"

pval_adjust_method
a string the indicates the method that is used to adjust the p-value for the multiple testing. It must match the options in p.adjust.methods. Default: "BH"

location_prior_df
the number of degrees of freedom used for the location prior. A large number (> 30) means that the prior is approximately Normal. Default: 3

max_iter
the maximum of iterations proDA() tries to converge to the hyper-parameter estimates. Default: 20

epsilon
if the remaining error is smaller than epsilon the model has converged. Default: 1e-3

return_fit
boolean that signals that in addition to the data.frame with the hypothesis test results, the fit from proDA() is returned. Default: FALSE

verbose
boolean that signals if the method prints informative messages. Default: FALSE

groups
a factor or character vector with that assignes the columns of X to different conditions. This parameter is only applicable for the F-test and must be specified if only a single matrix is provided.

Value

If return_fit == FALSE a data.frame is returned with the content that is described in test_diff.

If return_fit == TRUE a list is returned with two elements: fit with a reference to the object returned from proDA() and a test_result() with the data.frame returned from test_diff().

Details

The pd_row_t_test is not acutally doing a t-test, but rather a Wald test. But, as the two are closely related and term t-test is more widely understood, we choose to use that name.

See also

Examples

  data1 <- matrix(rnorm(10 * 3), nrow=10)
  data2 <- matrix(rnorm(10 * 4), nrow=10)
  data3 <- matrix(rnorm(10 * 2), nrow=10)

  # Comparing two datasets
  pd_row_t_test(data1, data2)

#>    name      pval  adj_pval        diff t_statistic        se df avg_abundance
#> 1     1 0.9717134 0.9718566 -0.01069420 -0.03726813 0.2869529  5  0.0879573189
#> 2     2 0.8789066 0.9718566  0.04573239  0.16031754 0.2852613  5  0.0677787793
#> 3     3 0.9654456 0.9718566 -0.01291091 -0.04553224 0.2835554  5  0.0711428256
#> 4     4 0.8156346 0.9718566 -0.07150091 -0.24576382 0.2909334  5  0.1114186933
#> 5     5        NA        NA          NA          NA        NA  5            NA
#> 6     6 0.9472993 0.9718566  0.01953687  0.06948186 0.2811795  5  0.0515140600
#> 7     7 0.9718566 0.9718566  0.01097429  0.03707932 0.2959681  5  0.1215490241
#> 8     8 0.7716905 0.9718566 -0.09102689 -0.30634159 0.2971418  5  0.1360018294
#> 9     9 0.7409478 0.9718566 -0.09575188 -0.34950692 0.2739628  5  0.0008229996
#> 10   10 0.8316209 0.9718566 -0.06195591 -0.22400225 0.2765861  5  0.0295052072
#>    n_approx n_obs
#> 1  7.000000     7
#> 2  7.000000     7
#> 3  7.000021     7
#> 4  7.000086     7
#> 5        NA     7
#> 6  7.000000     7
#> 7  7.000000     7
#> 8  7.000000     7
#> 9  7.000000     7
#> 10 7.000000     7

  # Comparing multiple datasets
  pd_row_f_test(data1, data2, data3)

#>    name      pval  adj_pval f_statistic df1 df2 avg_abundance n_approx n_obs
#> 1     1 0.9954189 0.9989838 0.004591606   2 Inf    0.07319044        9     9
#> 2     2        NA        NA          NA   2  NA            NA       NA     9
#> 3     3 0.9940074 0.9989838 0.006010640   2 Inf    0.06930352        9     9
#> 4     4 0.9894503 0.9989838 0.010605771   2 Inf    0.07716949        9     9
#> 5     5 0.9738045 0.9989838 0.026544761   2 Inf    0.07363987        9     9
#> 6     6 0.9989838 0.9989838 0.001016693   2 Inf    0.06051511        9     9
#> 7     7 0.9961467 0.9989838 0.003860781   2 Inf    0.07643498        9     9
#> 8     8 0.9496261 0.9989838 0.051686959   2 Inf    0.07461905        9     9
#> 9     9 0.9633848 0.9989838 0.037302319   2 Inf    0.05451577        9     9
#> 10   10 0.9873769 0.9989838 0.012703432   2 Inf    0.05264584        9     9

  # Alternative
  data_comb <- cbind(data1, data2, data3)
  pd_row_f_test(data_comb,
     groups = c(rep("A",3), rep("B", 4), rep("C", 2)))

#>    name      pval  adj_pval f_statistic df1 df2 avg_abundance n_approx n_obs
#> 1     1 0.9954189 0.9989838 0.004591606   2 Inf    0.07319044        9     9
#> 2     2        NA        NA          NA   2  NA            NA       NA     9
#> 3     3 0.9940074 0.9989838 0.006010640   2 Inf    0.06930352        9     9
#> 4     4 0.9894503 0.9989838 0.010605771   2 Inf    0.07716949        9     9
#> 5     5 0.9738045 0.9989838 0.026544761   2 Inf    0.07363987        9     9
#> 6     6 0.9989838 0.9989838 0.001016693   2 Inf    0.06051511        9     9
#> 7     7 0.9961467 0.9989838 0.003860781   2 Inf    0.07643498        9     9
#> 8     8 0.9496261 0.9989838 0.051686959   2 Inf    0.07461905        9     9
#> 9     9 0.9633848 0.9989838 0.037302319   2 Inf    0.05451577        9     9
#> 10   10 0.9873769 0.9989838 0.012703432   2 Inf    0.05264584        9     9

  # t.test, lm, pd_row_t_test, and pd_row_f_test are
  # approximately equivalent on fully observed data
  set.seed(1)
  x <- rnorm(5)
  y <- rnorm(5, mean=0.3)

  t.test(x, y)

#> 
#> 	Welch Two Sample t-test
#> 
#> data:  x and y
#> t = -0.58413, df = 7.1385, p-value = 0.5771
#> alternative hypothesis: true difference in means is not equal to 0
#> 95 percent confidence interval:
#>  -1.5391981  0.9274665
#> sample estimates:
#> mean of x mean of y 
#> 0.1292699 0.4351357 
#> 

  summary(lm(c(x, y) ~ cond,
             data = data.frame(cond = c(rep("x", 5),
                                        rep("y", 5)))))$coef[2,]

#> Warning: partial match of 'coef' to 'coefficients'

#>   Estimate Std. Error    t value   Pr(>|t|) 
#>  0.3058658  0.5236289  0.5841270  0.5752316 

  pd_row_t_test(matrix(x, nrow=1), matrix(y, nrow=1),
                moderate_location = FALSE,
                moderate_variance = FALSE)

#>   name      pval  adj_pval       diff t_statistic        se df avg_abundance
#> 1    1 0.5752316 0.5752316 -0.3058658   -0.584127 0.5236289  8     0.2822028
#>   n_approx n_obs
#> 1       10    10

  pd_row_f_test(matrix(x, nrow=1), matrix(y, nrow=1),
                moderate_location = FALSE,
                moderate_variance = FALSE)

#>   name      pval  adj_pval f_statistic df1 df2 avg_abundance n_approx n_obs
#> 1    1 0.5752316 0.5752316   0.3412044   1   8     0.2822028       10    10