Main function to determine the hyper and protein parameters

The function fits a linear probabilistic dropout model and infers the hyper-parameters for the location prior, the variance prior, and the dropout curves. In addition it infers for each protein the coefficients that best explain the observed data and the associated uncertainty.

proDA(data, design = ~1, col_data = NULL, reference_level = NULL,
  data_is_log_transformed = TRUE, moderate_location = TRUE,
  moderate_variance = TRUE, location_prior_df = 3,
  n_subsample = nrow(data), max_iter = 20, epsilon = 0.001,
  verbose = FALSE, ...)

Arguments

data
a matrix like object (matrix() or SummarizedExperiment()) with the one column per sample and one row per protein. Missing values should be coded NA.

design
a specification of the experimental design that is used to fit the linear model. It can be a model.matrix() with one row for each sample and one column for each coefficient. It can also be a formula with the entries refering to global objects, columns in the col_data argument or columns in the colData(data) if data is a SummarizedExperiment. Thirdly, it can be a vector that for each sample specifies the condition of that sample. Default: ~ 1, which means that all samples are treated as if they are in the same condition.

col_data
a data.frame with one row for each sample in data. Default: NULL

reference_level
a string that specifies which level in a factor coefficient is used for the intercept. Default: NULL

data_is_log_transformed
the raw intensities from mass spectrometry experiments have a linear mean-variance relation. This is undesirable and can be removed by working on the log scale. The easiest way to find out if the data is already log- transformed is to see if the intensities are in the range of 0 to 100 in which case they are transformed or if they rather are between 1e5 to 1e12, in which case they need to be transformed. Default: TRUE

moderate_location, moderate_variance
boolean values to indicate if the location and the variances are moderated. Default: TRUE

location_prior_df
the number of degrees of freedom used for the location prior. A large number (> 30) means that the prior is approximately Normal. Default: 3

n_subsample
the number of proteins that are used to estimate the hyper-parameter. Reducing this number can speed up the fitting, but also mean that the final estimate is less precise. By default all proteins are used. Default: nrow(data)

max_iter
the maximum of iterations proDA() tries to converge to the hyper-parameter estimates. Default: 20

epsilon
if the remaining error is smaller than epsilon the model has converged. Default: 1e-3

verbose
boolean that signals if the method prints informative messages. Default: FALSE

...
additional parameters for the construction of the 'proDAFit' object

Value

An object of class 'proDAFit'. The object contains information on the hyper-parameters and feature parameters, the convergence, the experimental design etc. Internally, it is a sub-class of SummarizedExperiment which means the object is subsettable.

Details

By default, the method is moderating the locations and the variance of each protein estimate. The variance moderation is fairly standard in high-throughput experiments and can boost the power to detect differentially abundant proteins. The location moderation is important to handle extreme cases where in one conditio a protein is not observed in any sample. In addition it can help to get more precise estimates of the difference between conditions. Unlike 'DESeq2', which moderates the coefficient estimates (ie. the "betas") to be centered around zero, 'proDA' penalizes predicted intensities that strain far from the other observed intensities.

Examples

# Quick start
set.seed(1)
library(proDA)
syn_data <- generate_synthetic_data(n_proteins = 10)
fit <- proDA(syn_data$Y, design = syn_data$groups)

#> Warning: Dropout curve estimation did not properly converge

fit

#> 	Parameters of the probabilistic dropout model
#> 
#> The dataset contains 6 samples and 10 proteins
#> 23.3% of the values are missing
#> 
#> Experimental design was specified using a design matrix (design(object)).
#> The model has successfully converged.
#> 
#> The inferred parameters are:
#> location_prior_mean:     19.7
#> location_prior_scale:    1.95
#> location_prior_df:       3
#> variance_prior_scale:    0.0179
#> variance_prior_df:       0.724
#> dropout_curve_position:  18.7, 18.1, 19.1, 19.1, ...
#> dropout_curve_scale:     -0.0001, -0.0001, -0.0001, -0.0001, ...

result_names(fit)

#> [1] "Condition_1" "Condition_2"

test_diff(fit, Condition_1 - Condition_2)

#>          name         pval    adj_pval         diff t_statistic         se
#> 1   protein_1 0.7060763154 0.996231662  0.072690682  0.40073981 0.18139122
#> 2   protein_2 0.4237058276 0.847411655  0.401437653  0.92315012 0.43485631
#> 3   protein_3 0.2120055631 0.634356389  0.747953323  1.44240629 0.51854552
#> 4   protein_4 0.6973046672 0.996231662 -0.148779812 -0.55760827 0.26681780
#> 5   protein_5 0.2537425556 0.634356389  0.382277102  1.29881328 0.29432799
#> 6   protein_6 0.9967733556 0.996773356 -0.507729191 -0.39999134 1.26935045
#> 7   protein_7 0.7969853297 0.996231662  0.198034364  0.28979781 0.68335356
#> 8   protein_8 0.1730855067 0.634356389 -0.159082184 -1.60343657 0.09921327
#> 9   protein_9 0.9764973789 0.996773356  0.003915554  0.03105014 0.12610425
#> 10 protein_10 0.0007393126 0.007393126 -1.158100457 -7.73846509 0.14965506
#>           df avg_abundance n_approx n_obs
#> 1  4.7276086      19.39852 6.000000     6
#> 2  3.0150835      18.99498 4.287475     4
#> 3  4.7276086      20.25723 6.000000     6
#> 4  0.7656805      18.18300 2.038072     2
#> 5  4.7276086      20.97573 6.000000     6
#> 6  0.0010000      17.86300 2.001000     0
#> 7  2.2045986      19.36016 3.476990     4
#> 8  4.7276086      20.63405 6.000000     6
#> 9  4.7276086      20.23926 6.000000     6
#> 10 4.7276086      22.08977 6.000000     6

# SummarizedExperiment
se <- generate_synthetic_data(n_proteins = 10,
                     return_summarized_experiment = TRUE)
se

#> class: SummarizedExperiment 
#> dim: 10 6 
#> metadata(0):
#> assays(2): abundances full_observations
#> rownames(10): protein_1 protein_2 ... protein_9 protein_10
#> rowData names(4): changed true_s2 true_Condition_1 true_Condition_2
#> colnames(6): Condition_1-1 Condition_1-2 ... Condition_2-2
#>   Condition_2-3
#> colData names(3): group true_dropout_curve_position
#>   true_dropout_curve_scale

proDA(se, design = ~ group)

#> Warning: Dropout curve estimation did not properly converge

#> 	Parameters of the probabilistic dropout model
#> 
#> The dataset contains 6 samples and 10 proteins
#> 18.3% of the values are missing
#> 
#> Experimental design: y~group
#> The model has successfully converged.
#> 
#> The inferred parameters are:
#> location_prior_mean:     20.7
#> location_prior_scale:    7.23
#> location_prior_df:       3
#> variance_prior_scale:    0.0538
#> variance_prior_df:       1.8
#> dropout_curve_position:  NA, 19.9, 19, 19.1, ...
#> dropout_curve_scale:     NA, -0.0001, -0.824, -0.0001, ...

# Design using model.matrix()
data_mat <- matrix(rnorm(5 * 10), nrow=10)
colnames(data_mat) <- paste0("sample", 1:5)
annotation_df <- data.frame(names = paste0("sample", 1:5),
                     condition = c("A", "A", "A", "B", "B"),
                     age = rnorm(5, mean=40, sd=10))

design_mat <- model.matrix(~ condition + age,
                           data=annotation_df)
design_mat

#>   (Intercept) conditionB      age
#> 1           1          0 24.85113
#> 2           1          0 44.34537
#> 3           1          0 34.80463
#> 4           1          1 31.65441
#> 5           1          1 32.43352
#> attr(,"assign")
#> [1] 0 1 2
#> attr(,"contrasts")
#> attr(,"contrasts")$condition
#> [1] "contr.treatment"
#> 

proDA(data_mat, design_mat, col_data = annotation_df)

#> 	Parameters of the probabilistic dropout model
#> 
#> The dataset contains 5 samples and 10 proteins
#> 0% of the values are missing
#> 
#> Experimental design was specified using a design matrix (design(object)).
#> The model has successfully converged.
#> 
#> The inferred parameters are:
#> location_prior_mean:     -0.00877
#> location_prior_scale:    0.657
#> location_prior_df:       3
#> variance_prior_scale:    0.798
#> variance_prior_df:       > 100
#> dropout_curve_position:  NA, NA, NA, NA, ...
#> dropout_curve_scale:     NA, NA, NA, NA, ...