Identify significant coefficients

The function is used to test coefficients of a 'proDAFit' object. It provides a Wald test to test individual coefficients and a likelihood ratio F-test to compare the original model with a reduced model. The result_names method provides a quick overview which coefficients are available for testing.

test_diff(fit, contrast, reduced_model = ~1,
  alternative = c("two.sided", "greater", "less"),
  pval_adjust_method = "BH", sort_by = NULL, decreasing = FALSE,
  n_max = Inf, verbose = FALSE)

# S4 method for proDAFit
result_names(fit)

Arguments

fit
an object of class 'proDAFit'. Usually, this is produced by calling proDA()

contrast
an expression or a string specifying which contrast is tested. It can be a single coefficient (to see the available options use result_names(fit)) or any linear combination of them. The contrast is always compared against zero. Thus, to find out if two coefficients differ use coef1 - coef2.

reduced_model
The fit with an alternative model (nested in the original model) is compared with the original model using an F-test. This is useful if not just an individual coefficient should be tested, but the effect of including a combination of covariates. If neither a contrast, nor a reduced_model is specified, by default a comparison with an intercept model is done. Default: ~ 1.

alternative
a string that decides how the hypothesis test is done. This parameter is only relevant for the t-test / contrast test. Default: "two.sided"

pval_adjust_method
a string the indicates the method that is used to adjust the p-value for the multiple testing. It must match the options in p.adjust.methods. Default: "BH"

sort_by
a string that specifies the column that is used to sort the resulting data.frame. Default: NULL which means the result is sorted by the order of the input matrix.

decreasing
a boolean to indicate if the order is reversed. Default: FALSE

n_max
the maximum number of rows returned by the method. Default: Inf

verbose
boolean that signals if the method prints informative messages. Default: FALSE.

Value

Both functions return a data.frame with one row per protein with the key parameters of the statistical test.

The Wald test, which can considered equivalent to a t-test, returns a data.frame with the following columns:

name

the name of the protein, extracted from the rowname of the input matrix

pval

the p-value of the statistical test

adj_pval

the multiple testing adjusted p-value

diff

the difference that particular coefficient makes. In differential expression analysis this value is also called log fold change, which is equivalent to the difference on the log scale.

t_statistic

the diff divided by the standard error se

se

the standard error associated with the diff

df

the degrees of freedom, which describe the amount of available information for estimating the se. They are the sum of the number of samples the protein was observed in, the amount of information contained in the missing values, and the estimated df from the variance prior.

avg_abundance

the estimate of the average abundance of the protein across all samples.

n_approx

the approximated information available for estimating the protein features, expressed as multiple of the information contained in one observed value.

n_obs

the number of samples a protein was observed in

The F-test returns a data.frame with the following columns

name

the name of the protein, extracted from the rowname of the input matrix

pval

the p-value of the statistical test

adj_pval

the multiple testing adjusted p-value

f_statistic

the ratio of difference of normalized deviances from original model and the reduced model, divided by the standard deviation.

df1

the difference of the number of coefficients in the original model and the number of coefficients in the reduced model

df2

the degrees of freedom, which describe the amount of available information for estimating the se. They are the sum of the number of samples the protein was observed in, the amount of information contained in the missing values, and the estimated df from the variance prior.

avg_abundance

the estimate of the average abundance of the protein across all samples.

n_approx

the information available for estimating the protein features, expressed as multiple of the information contained in one observed value.

n_obs

the number of samples a protein was observed in

Details

To test if coefficient is different from zero with a Wald test use the contrast function argument. To test if two models differ with an F-test use the reduced_model argument. Depending on the test that is conducted, the functions returns slightly different data.frames.

The function is designed to follow the principles of the base R test functions (ie. t.test and wilcox.test) and the functions designed for collecting the results of high-throughput testing (ie. limma::topTable and DESeq2::results).

See also

Examples

  # "t-test"
  syn_data <- generate_synthetic_data(n_proteins = 10)
  fit <- proDA(syn_data$Y, design = syn_data$groups)
  result_names(fit)

#> [1] "Condition_1" "Condition_2"

  test_diff(fit, Condition_1 - Condition_2)

#>          name      pval  adj_pval         diff  t_statistic        se  df
#> 1   protein_1 0.9935739 0.9935739 -0.003683657 -0.008053976 0.4573712 Inf
#> 2   protein_2 0.6532176 0.9935739  0.205495368  0.449296694 0.4573712 Inf
#> 3   protein_3 0.6769552 0.9935739 -0.190550721 -0.416621601 0.4573712 Inf
#> 4   protein_4 0.6459248 0.9935739 -0.210130465 -0.459430907 0.4573712 Inf
#> 5   protein_5 0.9844568 0.9935739  0.008910395  0.019481759 0.4573712 Inf
#> 6   protein_6 0.8250244 0.9935739 -0.101118984 -0.221087345 0.4573712 Inf
#> 7   protein_7 0.5395040 0.9935739 -0.280626156 -0.613563241 0.4573712 Inf
#> 8   protein_8 0.8453393 0.9935739  0.089218727  0.195068528 0.4573712 Inf
#> 9   protein_9 0.3992664 0.9935739 -0.385532912 -0.842932200 0.4573712 Inf
#> 10 protein_10 0.2652593 0.9935739  0.509532711  1.114046339 0.4573712 Inf
#>    avg_abundance n_approx n_obs
#> 1       21.08237 6.000006     6
#> 2       20.98692 6.000011     6
#> 3       20.07388 6.000000     6
#> 4       19.45017 4.986517     5
#> 5       19.24395 3.812491     4
#> 6       20.93682 6.000009     6
#> 7       18.51581 2.723117     2
#> 8       20.25957 4.906474     5
#> 9       19.21564 3.773366     4
#> 10      18.20284 2.653161     2

  suppressPackageStartupMessages(library(SummarizedExperiment))
  se <- generate_synthetic_data(n_proteins = 10,
                                n_conditions = 3,
                                return_summarized_experiment = TRUE)
  colData(se)$age <- rnorm(9, mean=45, sd=5)
  colData(se)

#> DataFrame with 9 rows and 4 columns
#>                     group true_dropout_curve_position true_dropout_curve_scale
#>                  <factor>                   <numeric>                <numeric>
#> Condition_1-1 Condition_1                        18.5                     -1.2
#> Condition_1-2 Condition_1                        18.5                     -1.2
#> Condition_1-3 Condition_1                        18.5                     -1.2
#> Condition_2-1 Condition_2                        18.5                     -1.2
#> Condition_2-2 Condition_2                        18.5                     -1.2
#> Condition_2-3 Condition_2                        18.5                     -1.2
#> Condition_3-1 Condition_3                        18.5                     -1.2
#> Condition_3-2 Condition_3                        18.5                     -1.2
#> Condition_3-3 Condition_3                        18.5                     -1.2
#>                            age
#>                      <numeric>
#> Condition_1-1 53.8179317387294
#> Condition_1-2 48.4480011096705
#> Condition_1-3 39.4962967778804
#> Condition_2-1 48.5725467844891
#> Condition_2-2 43.7676484153299
#> Condition_2-3  43.401069170364
#> Condition_3-1 51.8132214647735
#> Condition_3-2 38.8605870500866
#> Condition_3-3 42.4439038362476

  fit <- proDA(se, design = ~ group + age)

#> Warning: Dropout curve estimation did not properly converge

  result_names(fit)

#> [1] "Intercept"        "groupCondition_2" "groupCondition_3" "age"             

  test_diff(fit, "groupCondition_2",
            n_max = 3, sort_by = "pval")

#>          name         pval    adj_pval       diff t_statistic        se
#> 10 protein_10 0.0000177626 0.000177626 -4.4503631 -10.3378643 0.4304915
#> 2   protein_2 0.0922230978 0.461115489  0.8318491   1.9510714 0.4263550
#> 5   protein_5 0.3788492204 0.992013396  0.4142066   0.9557837 0.4333686
#>          df avg_abundance n_approx n_obs
#> 10 6.966178      25.79399 9.000000     9
#> 2  6.966178      20.73585 9.000000     9
#> 5  5.562959      19.63621 7.596781     8

  # F-test
  test_diff(fit, reduced_model = ~ group)

#>          name      pval  adj_pval f_statistic df1      df2 avg_abundance
#> 1   protein_1 1.0000000 1.0000000 -2.09851282   1 0.001000      17.53187
#> 2   protein_2 0.0721679 0.7216790  4.48419304   1 6.966178      20.73585
#> 3   protein_3 0.1642050 0.8210248  2.51812436   1 5.932027      20.19210
#> 4   protein_4 0.5482641 1.0000000  0.39798095   1 6.966178      22.43375
#> 5   protein_5 1.0000000 1.0000000 -0.01676887   1 5.562959      19.63621
#> 6   protein_6 0.8304182 1.0000000  0.04944042   1 6.966178      19.23689
#> 7   protein_7 0.8659416 1.0000000  0.03957787   1 1.496228      17.79062
#> 8   protein_8 1.0000000 1.0000000 -0.24959699   1 2.513792      18.90239
#> 9   protein_9 1.0000000 1.0000000 -0.01927298   1 4.902852      19.22842
#> 10 protein_10 0.6370960 1.0000000  0.24315976   1 6.966178      25.79399
#>    n_approx n_obs
#> 1  4.001000     1
#> 2  9.000000     9
#> 3  7.965850     8
#> 4  9.000000     9
#> 5  7.596781     8
#> 6  9.000000     9
#> 7  3.530050     3
#> 8  4.547614     4
#> 9  6.936674     7
#> 10 9.000000     9